Extract nested subclasses of Command
Currently, Command nests all the subtypes of command as part of itself.
This allows the abstract Command Class to conveniently return its
sub-types using the factory method.

However, this was done when the classes have not been properly packaged,
where having all the classes be in 1 folder would have bloated the
project environment significantly. The tradeoff for this would be
causing the Command class itself to bloat unnecessarily.

Therefore, we choose to extract the subclasses inside Command into the
trackerbot.command package.

To maintain the original design philosophy of having only Command be
able to access these classes, we set the relevant access modifiers to
package-private.
