+++ src/main/java/trackerbot/command/ExitCommand.java
@@ -1,0 +1,13 @@
+package trackerbot.command;
+
+import trackerbot.gui.UiHandler;
+import trackerbot.task.TaskList;
+
+class ExitCommand extends Command {
+    ExitCommand() {}
+
+    @Override
+    public void execute(TaskList tasks, UiHandler uiHandler) {
+        uiHandler.exitApp();
+    }
+}
