+++ src/main/java/trackerbot/command/FindCommand.java
@@ -1,0 +1,17 @@
+package trackerbot.command;
+
+import trackerbot.gui.UiHandler;
+import trackerbot.task.TaskList;
+
+class FindCommand extends Command {
+    private final String commandFields;
+
+    FindCommand(String commandFields) {
+        this.commandFields = commandFields;
+    }
+
+    @Override
+    public void execute(TaskList tasks, UiHandler uiHandler) {
+        uiHandler.setMessage(tasks.findAll(commandFields));
+    }
+}
