+++ src/main/java/trackerbot/command/ListCommand.java
@@ -1,0 +1,13 @@
+package trackerbot.command;
+
+import trackerbot.gui.UiHandler;
+import trackerbot.task.TaskList;
+
+class ListCommand extends Command {
+    ListCommand() {}
+
+    @Override
+    public void execute(TaskList tasks, UiHandler uiHandler) {
+        uiHandler.setMessage(tasks.list());
+    }
+}
