+++ src/main/java/trackerbot/command/UnknownCommand.java
@@ -1,0 +1,14 @@
+package trackerbot.command;
+
+import trackerbot.exception.TrackerBotException;
+import trackerbot.gui.UiHandler;
+import trackerbot.task.TaskList;
+
+class UnknownCommand extends Command {
+    UnknownCommand() {}
+
+    @Override
+    public void execute(TaskList tasks, UiHandler uiHandler) throws TrackerBotException {
+        throw new TrackerBotException("Unrecognised Command Type. Try another?");
+    }
+}
